requirements.txt
flask
ultralytics
opencv-python
transformers
Pillow